requirements.txt
opencv-python==4.10.0
deep_sort_realtime==1.3.2
filterpy==1.4.5
lap==0.4.0
matplotlib==3.9.2
numpy==1.25.2
scipy==1.14.1
skimage==0.0
torch==2.5.0
ultralytics==8.3.18